+++ messaging/coap/coap.h
@@ -106,7 +106,7 @@ typedef struct
   uint8_t token_len;
   uint8_t token[COAP_TOKEN_LEN];
 
-  uint8_t options[COAP_OPTION_SIZE1 / (int)OPTION_MAP_SIZE +
+  uint8_t options[COAP_OPTION_SIZE1 / OPTION_MAP_SIZE +
                   1]; /* bitmap to check if option is set */
 
   uint16_t content_format; /* parse options once and store; allows setting
